Refactor tests to run queries on non authenticated cluster Authenticated tests will confirm cluster can authenticate on basic operation Added authentication properties to auto configuration and adoc
