+++ auto-configurations/vector-stores/spring-ai-autoconfigure-vector-store-gemfire/src/main/java/org/springframework/ai/vectorstore/gemfire/autoconfigure/GemFireConnectionDetails.java
@@ -29,4 +29,16 @@ public interface GemFireConnectionDetails extends ConnectionDetails {
 
 	int getPort();
 
+	default String getUsername() {
+		return null;
+	}
+
+	default String getPassword() {
+		return null;
+	}
+
+	default String getToken() {
+		return null;
+	}
+
 }
