+++ auto-configurations/vector-stores/spring-ai-autoconfigure-vector-store-gemfire/src/test/java/org/springframework/ai/vectorstore/gemfire/autoconfigure/GemFireVectorStorePropertiesTests.java
@@ -38,6 +38,9 @@ class GemFireVectorStorePropertiesTests {
 		assertThat(props.getMaxConnections()).isEqualTo(GemFireVectorStore.DEFAULT_MAX_CONNECTIONS);
 		assertThat(props.getFields()).isEqualTo(GemFireVectorStore.DEFAULT_FIELDS);
 		assertThat(props.getBuckets()).isEqualTo(GemFireVectorStore.DEFAULT_BUCKETS);
+		assertThat(props.getUsername()).isNull();
+		assertThat(props.getPassword()).isNull();
+		assertThat(props.getToken()).isNull();
 	}
 
 	@Test
@@ -50,6 +53,9 @@ class GemFireVectorStorePropertiesTests {
 		props.setMaxConnections(10);
 		props.setFields(new String[] { "test" });
 		props.setBuckets(10);
+		props.setUsername("username");
+		props.setPassword("password");
+		props.setToken("token");
 
 		assertThat(props.getIndexName()).isEqualTo("spring-ai-index");
 		assertThat(props.getHost()).isEqualTo("localhost");
@@ -58,6 +64,9 @@ class GemFireVectorStorePropertiesTests {
 		assertThat(props.getMaxConnections()).isEqualTo(10);
 		assertThat(props.getFields()).isEqualTo(new String[] { "test" });
 		assertThat(props.getBuckets()).isEqualTo(10);
+		assertThat(props.getUsername()).isEqualTo("username");
+		assertThat(props.getPassword()).isEqualTo("password");
+		assertThat(props.getToken()).isEqualTo("token");
 
 	}
