+++ vector-stores/spring-ai-gemfire-store/src/test/java/org/springframework/ai/vectorstore/gemfire/GemFireWithTokenAuthenticationVectorStoreIT.java
@@ -16,12 +16,7 @@
 
 package org.springframework.ai.vectorstore.gemfire;
 
-import com.github.dockerjava.api.model.ExposedPort;
-import com.github.dockerjava.api.model.PortBinding;
-import com.github.dockerjava.api.model.Ports;
-import com.vmware.gemfire.testcontainers.GemFireCluster;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Disabled;
 
 import org.springframework.ai.embedding.EmbeddingModel;
 import org.springframework.ai.transformers.TransformersEmbeddingModel;
@@ -32,34 +27,8 @@ import org.springframework.context.annotation.Bean;
 /**
  * @author Jason Huynh
  */
-public class GemFireWithTokenAuthenticationVectorStoreIT extends GemFireVectorStoreBaseIT {
-
-	@AfterAll
-	public static void stopGemFireCluster() {
-		gemFireCluster.close();
-	}
-
-	@BeforeAll
-	public static void startGemFireCluster() {
-		Ports.Binding hostPort = Ports.Binding.bindPort(HTTP_SERVICE_PORT);
-		ExposedPort exposedPort = new ExposedPort(HTTP_SERVICE_PORT);
-		PortBinding mappedPort = new PortBinding(hostPort, exposedPort);
-		gemFireCluster = new GemFireCluster(GemFireImage.DEFAULT_IMAGE, LOCATOR_COUNT, SERVER_COUNT);
-		gemFireCluster.withConfiguration(GemFireCluster.SERVER_GLOB,
-				container -> container.withExposedPorts(HTTP_SERVICE_PORT)
-					.withCreateContainerCmdModifier(cmd -> cmd.getHostConfig().withPortBindings(mappedPort)));
-		gemFireCluster.withGemFireProperty(GemFireCluster.SERVER_GLOB, "http-service-port",
-				Integer.toString(HTTP_SERVICE_PORT));
-		gemFireCluster.withGemFireProperty(GemFireCluster.ALL_GLOB, "security-manager",
-				"org.apache.geode.examples.SimpleSecurityManager");
-
-		gemFireCluster.withGemFireProperty(GemFireCluster.ALL_GLOB, "security-username", "clusterManage");
-		gemFireCluster.withGemFireProperty(GemFireCluster.ALL_GLOB, "security-password", "clusterManage");
-		gemFireCluster.acceptLicense().start();
-
-		System.setProperty("spring.data.gemfire.pool.locators",
-				String.format("localhost[%d]", gemFireCluster.getLocatorPort()));
-	}
+@Disabled
+public class GemFireWithTokenAuthenticationVectorStoreIT extends GemFireVectorStoreAuthenticationBaseIT {
 
 	@Override
 	Class getTestApplicationClass() {
